Refactored tool
